+++ tests/framework/db/ActiveRecordTest.php
@@ -527,6 +527,7 @@ class ActiveRecordTest extends DatabaseTestCase
         $this->assertTrue($customer->orders2[0]->customer2 === $customer);
         $this->assertTrue($customer->orders2[1]->customer2 === $customer);
         // ad-hoc lazy loading
+        // NB: this batch of asserts probably does not do what was intended ($customer->orders2[0] vs $orders[0])
         $customer = Customer::findOne(2);
         $orders = $customer->getOrders2()->all();
         $this->assertTrue(count($orders) === 2);
@@ -563,6 +564,32 @@ class ActiveRecordTest extends DatabaseTestCase
         $this->assertTrue($orders[0]['customer2']['orders2'][1]['id'] === $orders[1]['id']);
         $this->assertTrue($orders[1]['customer2']['orders2'][0]['id'] === $orders[0]['id']);
         $this->assertTrue($orders[1]['customer2']['orders2'][1]['id'] === $orders[1]['id']);
+    }
+
+    public function testInverseOfDynamic()
+    {
+        $customer = Customer::findOne(1);
+        
+        // request the inverseOf relation without explicitly (eagerly) loading it
+        $orders2 = $customer->getOrders2()->all();
+        $this->assertSame($customer, $orders2[0]->customer2);
+        
+        $orders2 = $customer->getOrders2()->one();
+        $this->assertSame($customer, $orders2->customer2);
+        
+        // request the inverseOf relation while also explicitly eager loading it (while possible, this is of course redundant)
+        $orders2 = $customer->getOrders2()->with('customer2')->all();
+        $this->assertSame($customer, $orders2[0]->customer2);
+        
+        $orders2 = $customer->getOrders2()->with('customer2')->one();
+        $this->assertSame($customer, $orders2->customer2);
+        
+        // request the inverseOf relation as array
+        $orders2 = $customer->getOrders2()->asArray()->all();
+        $this->assertEquals($customer['id'], $orders2[0]['customer2']['id']);
+        
+        $orders2 = $customer->getOrders2()->asArray()->one();
+        $this->assertEquals($customer['id'], $orders2['customer2']['id']);
     }
 
     public function testDefaultValues()